added various trivial improvements and fixes

+++ libraries/include/forsyth_too.h
@@ -81,8 +81,8 @@ namespace Forsyth
                 {
                     assert ( cachePosition < vertexCacheSize );
                     // Points for being high in the cache.
-                    const float scaler = 1.0f / ( vertexCacheSize - 3 );
-                    score = 1.0f - ( cachePosition - 3 ) * scaler;
+                    const float scaler = 1.0f / static_cast<float>( vertexCacheSize - 3 );
+                    score = 1.0f - static_cast<float>( cachePosition - 3 ) * scaler;
                     score = powf ( score, FindVertexScore_CacheDecayPower );
                 }
             }
